+++ packages/ketcher-core/src/chemistry/constants/Elements.ts
