+++ packages/ketcher-core/src/chemistry/constants/index.ts
@@ -1,4 +1,4 @@
 export * from './ElementColor'
-export * from './ElementsMap'
+export * from './Elements'
 export * from './element.types'
 export * from './genericsTree'
